pyproject.toml
[tool.poetry]
name = "ratchet-sm"
version = "0.0.1"
description = "A provider-agnostic state machine for handling, normalizing, and recovering structured LLM outputs."
authors = ["sonic182 <johander1822@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-frontmatter = "*"
pyyaml = "*"
pydantic = { version = ">=2.0", optional = true }

[tool.poetry.extras]
pydantic = ["pydantic"]

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
ruff = "*"
mypy = "*"
pydantic = ">=2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["S101"]

[tool.mypy]
python_version = "3.10"
strict = true

[[tool.mypy.overrides]]
module = ["frontmatter", "yaml", "pydantic"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=ratchet_sm --cov-report=term-missing"
